Add Jest and CircleCI configuration (#55)
* Add configuration for Jest

We are configuring Jest via `jest.config.js`, rather than through
CLI arguments or a block in `package.json`, as this seems like the
most straightforward configuration approach to maintain.

Things of note:

- We are collecting code coverage stats
- Alongside the default reporter, we are using the jest-junit
  reporter as this is the reporter that CircleCI supports
  https://circleci.com/docs/2.0/collect-test-data/#jest

* Remove absolute filepaths from Jest snapshots

There were previously three tests that generated snapshots containing
real absolute filepaths to JSON configuration files. Unfortunately
these tests will fail in a continuous integration environment
(e.g. CircleCI) as its filesystem won't be organised in the same way
as my development machine (!), resulting in the real absolute filepaths
being different.

In order to fix this issue I've updated these three tests to mock
the loading of config from a fake URL (using `nock`), meaning that the
snapshots generated should be consistent regardless of the environment
that they are generated in.

Additionally, the `collectMockCalls` helper is no longer serializing
the object it constructs, as Jest handles that automatically for us.
This has made our snapshot files a lot more readable and they should be
easier to diff in future.

* Add basic CircleCI configuration

* Pin node to 10.15.0 for consistency